refactor: 팀원들이랑 패키지 구조 맞춤
